Task_1_2_2/src/test/java/ru/nsu/fitkulin/HashTableTest.java
package ru.nsu.fitkulin;

import static org.junit.jupiter.api.Assertions.assertEquals;

import java.util.Iterator;
import java.util.Map.Entry;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

class HashTableTest {
    private HashTable<String, Integer> hashTable;

    @BeforeEach
    public void setUp() {
        hashTable = new HashTable<>();
    }

    @Test
    public void testHashTableOperations() {
        hashTable.put("one", 1);
        hashTable.put("two", 2);
        hashTable.put("three", 3);

        assertEquals(1, hashTable.get("one"));
        assertEquals(2, hashTable.get("two"));
        assertEquals(3, hashTable.get("three"));

        Iterator<Entry<String, Integer>> iterator = hashTable.iterator();
        int count = 0;
        while (iterator.hasNext()) {
            count++;
            iterator.next();
        }
        assertEquals(3, count);


        hashTable.removeKey("two");
        assertEquals(2, hashTable.getSize());
    }

    @Test
    public void testEmptyHashTable() {
        assertEquals(0, hashTable.getSize());
    }
}